+++ cronjobs/__after_compose.py
@@ -14,6 +14,8 @@ def after_compose(config, settings, yml, globals):
         current_dir.parent / "common_snippets" / "set_docker_group.sh",
         current_dir / "bin" / "set_docker_group.sh",
     )
-    src = current_dir.parent / "wodoo" / "src"
-    dest = current_dir.parent / "cronjobs" / "wodoo_src"
-    globals["tools"].sync_folder(src, dest, excludes=[".git", ".pyc", "__pycache__"])
+    src = current_dir.parent / "zodoo" / "src"
+    dest = current_dir.parent / "cronjobs" / "zodoo_src"
+    globals["tools"].sync_folder(
+        src, dest, excludes=[".git", ".pyc", "__pycache__"]
+    )