+++ cronjobshell/__after_compose.py
@@ -9,6 +9,6 @@ dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 
 
 def after_compose(config, settings, yml, globals):
-    src = current_dir.parent / "wodoo" / "src"
-    dest = current_dir.parent / "cronjobs" / "wodoo_src"
+    src = current_dir.parent / "zodoo" / "src"
+    dest = current_dir.parent / "cronjobs" / "zodoo_src"
     globals["tools"].sync_folder(src, dest, excludes=[".git"])