+++ zodoo/src/gimera_src/gimera/__init__.py
@@ -1,0 +1,16 @@
+import click
+
+
+class RuntimeState:
+    mode = None
+    flags = {}
+
+
+runtime_state = {"temppaths": {}}
+
+from . import consts
+from . import gimera
+from . import repo
+from . import gitcommands
+from . import tools
+from . import snapshot